% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\docType{class}
\name{DecisionTree-class}
\alias{DecisionTree-class}
\alias{DecisionTree}
\title{Decision Tree Model}
\value{
DecisionTree object
}
\description{
Tree-based classification/regression algorithm.
}

