% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticRegression.R
\docType{class}
\name{LogisticRegression-class}
\alias{LogisticRegression-class}
\alias{LogisticRegression}
\title{Logistic Regression Model}
\value{
LogisticRegression object
}
\description{
Binary classification logistic regression.
}

