% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot Confusion Matrix}
\usage{
plot_confusion_matrix(cm, normalize = TRUE)
}
\arguments{
\item{cm}{confusion matrix}

\item{normalize}{Normlize}
}
\value{
plot
}
\description{
Visualizes confusion matrix.
}
