% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{precision_score}
\alias{precision_score}
\title{Precision Score}
\usage{
precision_score(y_true, y_pred, positive = NULL)
}
\arguments{
\item{y_true}{true labels}

\item{y_pred}{predicted labels}

\item{positive}{positive class label}
}
\value{
numeric precision
}
\description{
Computes precision metric.
}
