% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{predict_linear_model}
\alias{predict_linear_model}
\title{Predict Linear Model}
\usage{
predict_linear_model(model, X)
}
\arguments{
\item{model}{trained model}

\item{X}{matrix}
}
\value{
numeric vector
}
\description{
Predict values using trained linear model.
}
