% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{recall_score}
\alias{recall_score}
\title{Recall Score}
\usage{
recall_score(y_true, y_pred, positive = NULL)
}
\arguments{
\item{y_true}{true labels}

\item{y_pred}{predicted labels}

\item{positive}{positive class label}
}
\value{
numeric recall
}
\description{
Computes recall metric.
}
