% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeRegression.R
\docType{class}
\name{RidgeRegression-class}
\alias{RidgeRegression-class}
\alias{RidgeRegression}
\title{Ridge Regression Model}
\value{
RidgeRegression object
}
\description{
Linear regression with L2 regularization.
}

