% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{VectorForgeML-package}
\alias{VectorForgeML}
\alias{VectorForgeML-package}
\title{VectorForgeML: High-Performance ML Framework}
\description{
Fast machine learning models implemented in C++.
}
\keyword{internal}
